wip: refactor

Took 9 minutes
