+++ views/head.py
@@ -1,6 +1,11 @@
 import logging
 
 import streamlit as st
+from streamlit.commands.page_config import (
+    REPORT_A_BUG_KEY,
+    ABOUT_KEY,
+    GET_HELP_KEY,
+)
 
 import public
 
@@ -28,6 +33,25 @@ def head():
 
 
 def head_v3():
+    st.set_page_config(
+        page_title="Decenter AI",
+        page_icon="static/favicon.ico",
+        layout="centered",
+        menu_items={
+            REPORT_A_BUG_KEY: "https://github.com/DeCenter-AI/decenter-ai.streamlit.app/issues/new/choose",
+            ABOUT_KEY: "https://app.pitch.com/app/dashboard/0ba0eb40-0ffc-4970-91a5-64cec23d3457",
+            GET_HELP_KEY: "https://github.com/DeCenter-AI/decenter-ai.streamlit.app/issues/new/choose",
+        },
+    )
+    st.sidebar.header("v3")
+
+    st.markdown(public.button_styles_css, unsafe_allow_html=True)
+
+    st.sidebar.markdown(
+        public.report_request_buttons_html,
+        unsafe_allow_html=True,
+    )
+
     st.write(public.index_css, unsafe_allow_html=True)
 
     col1, col2, col3 = st.columns([1, 2, 1])
